+++ src/CaptainHook.Common/Authentication/AuthenticationConfig.cs
@@ -1,4 +1,4 @@
-﻿namespace CaptainHook.Common
+﻿namespace CaptainHook.Common.Authentication
 {
     public abstract class AuthenticationConfig
     {
